Rename Edit Controllers
