+++ src/main/resources/bundles/strings.properties
@@ -1,10 +1,10 @@
 car-table-view-fxml=/fxml/CarTableView.fxml
-edit-car-view-fxml=/fxml/EditCarView.fxml
+car-edit-view-fxml=/fxml/CarEditView.fxml
 
 customer-table-view-fxml=/fxml/CustomerTableView.fxml
-edit-customer-view-fxml=/fxml/EditCustomerView.fxml
+customer-edit-view-fxml=/fxml/CustomerEditView.fxml
 
 employee-table-view-fxml=/fxml/EmployeeTableView.fxml
-edit-employee-view-fxml=/fxml/EditEmployeeView.fxml
+employee-edit-view-fxml=/fxml/EmployeeEditView.fxml
 
 rental-table-view-fxml=/fxml/RentalTableView.fxml
