+++ src/main/resources/fxml/CarEditView.fxml
@@ -1,0 +1,122 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.collections.FXCollections?>
+<?import javafx.geometry.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import java.util.HashMap?>
+<BorderPane xmlns="http://javafx.com/javafx"
+	xmlns:fx="http://javafx.com/fxml"
+	fx:controller="de.htwsaar.prog3.carrental.controller.CarEditViewController">
+	<top>
+		<BorderPane prefHeight="25.0" prefWidth="1442.0" BorderPane.alignment="CENTER">
+			<left>
+				<GridPane alignment="center" hgap="10" vgap="10">
+					<Label text="%car-edit-title-label" GridPane.columnIndex="0" GridPane.rowIndex="0" />
+				</GridPane>
+			</left>
+			<padding>
+				<Insets left="12.5" right="5.0" top="12.0" />
+			</padding>
+			<BorderPane.margin>
+				<Insets left="12.5" right="5.0" top="12.0" />
+			</BorderPane.margin>
+		</BorderPane>
+	</top>
+	<center>
+		<GridPane alignment="center" hgap="10" vgap="10">
+			<Label text="%car-brand-label" GridPane.columnIndex="0" GridPane.rowIndex="0" />
+			<TextField fx:id="brandTextField" GridPane.columnIndex="0" GridPane.rowIndex="1" />
+			<Label text="%car-model-label" GridPane.columnIndex="1" GridPane.rowIndex="0" />
+			<TextField fx:id="modelTextField" GridPane.columnIndex="1" GridPane.rowIndex="1" />
+			<Label text="%car-category-label" GridPane.columnIndex="2" GridPane.rowIndex="0" />
+			<TextField fx:id="categoryTextField" GridPane.columnIndex="2" GridPane.rowIndex="1" />
+			<Label text="%car-color-label" GridPane.columnIndex="3" GridPane.rowIndex="0" />
+			<TextField fx:id="colorTextField" GridPane.columnIndex="3" GridPane.rowIndex="1" />
+			<Label text="%car-construction-year-label" GridPane.columnIndex="4" GridPane.rowIndex="0" />
+			<TextField fx:id="constructionYearTextField" GridPane.columnIndex="4"
+				GridPane.rowIndex="1" />
+			<Label text="%car-driven-distance-label" GridPane.columnIndex="5" GridPane.rowIndex="0" />
+			<TextField fx:id="drivenDistanceTextField" GridPane.columnIndex="5"
+				GridPane.rowIndex="1" />
+			<Label text="%car-gearbox-label" GridPane.columnIndex="0" GridPane.rowIndex="2" />
+			<fx:define>
+				<HashMap fx:id="gearboxBundle" manual="%car-gearbox-manual-label"
+                         semiautomatic="%car-gearbox-semi-automatic-label" automatic="%car-gearbox-automatic-label" />
+			</fx:define>
+			<ChoiceBox fx:id="gearBoxChoiceBox" GridPane.columnIndex="0" GridPane.rowIndex="3"
+				GridPane.columnSpan="2">
+				<items>
+					<FXCollections fx:factory="observableArrayList">
+						<fx:reference source="gearboxBundle.manual" />
+						<fx:reference source="gearboxBundle.semiautomatic" />
+						<fx:reference source="gearboxBundle.automatic" />
+					</FXCollections>
+				</items>
+			</ChoiceBox>
+			<Label text="%car-horsepower-label" GridPane.columnIndex="1" GridPane.rowIndex="2" />
+			<TextField fx:id="horsePowerTextField" GridPane.columnIndex="1" GridPane.rowIndex="3" />
+			<Label text="%car-fuel-label" GridPane.columnIndex="2" GridPane.rowIndex="2" />
+			<fx:define>
+				<HashMap fx:id="fuelBundle" gasoline="%car-fuel-gasoline-label"
+					diesel="%car-fuel-diesel-label" electric="%car-fuel-electric-label"
+					hybrid="%car-fuel-hybrid-label" />
+			</fx:define>
+			<ChoiceBox fx:id="fuelChoiceBox" GridPane.columnIndex="2" GridPane.rowIndex="3">
+				<items>
+					<FXCollections fx:factory="observableArrayList">
+						<fx:reference source="fuelBundle.gasoline" />
+						<fx:reference source="fuelBundle.diesel" />
+						<fx:reference source="fuelBundle.electric" />
+						<fx:reference source="fuelBundle.hybrid" />
+					</FXCollections>
+				</items>
+			</ChoiceBox>
+			<Label text="%car-door-count-label" GridPane.columnIndex="3" GridPane.rowIndex="2" />
+			<TextField fx:id="doorCountTextField" GridPane.columnIndex="3" GridPane.rowIndex="3" />
+			<Label text="%car-tires-label" GridPane.columnIndex="4" GridPane.rowIndex="2" />
+			<TextField fx:id="tiresTextField" GridPane.columnIndex="4" GridPane.rowIndex="3" />
+			<Label text="%car-next-inspection-label" GridPane.columnIndex="5" GridPane.rowIndex="2" />
+			<TextField fx:id="nextInspectionTextField" GridPane.columnIndex="5"
+				GridPane.rowIndex="3" />
+			<Label text="%car-vin-label" GridPane.columnIndex="0" GridPane.rowIndex="4" />
+			<TextField fx:id="vinTextField" GridPane.columnIndex="0" GridPane.rowIndex="5" />
+			<Label text="%car-equipment-label" GridPane.columnIndex="0" GridPane.rowIndex="6" />
+			<TextField fx:id="equipmentTextField" GridPane.columnIndex="0" GridPane.rowIndex="7"
+				GridPane.columnSpan="6" />
+			<Label text="%car-defects-label" GridPane.columnIndex="0" GridPane.rowIndex="8" />
+			<TextField fx:id="defectsTextField" GridPane.columnIndex="0" GridPane.rowIndex="9"
+				GridPane.columnSpan="6" />
+			<Label text="%car-licence-number-label" GridPane.columnIndex="0" GridPane.rowIndex="10" />
+			<TextField fx:id="licenceNumberTextField" GridPane.columnIndex="0"
+				GridPane.rowIndex="11" />
+			<Label text="%car-parking-lot-label" GridPane.columnIndex="1" GridPane.rowIndex="10" />
+			<TextField fx:id="parkingLotTextField" GridPane.columnIndex="1" GridPane.rowIndex="11" />
+			<Label text="%car-daily-rate-label" GridPane.columnIndex="2" GridPane.rowIndex="10" />
+			<TextField fx:id="dailyRateTextField" GridPane.columnIndex="2" GridPane.rowIndex="11" />
+			<padding>
+				<Insets top="25" right="25" bottom="25" left="25" />
+			</padding>
+		</GridPane>
+	</center>
+	<bottom>
+		<BorderPane prefHeight="25.0" prefWidth="1442.0" BorderPane.alignment="CENTER">
+			<right>
+				<GridPane alignment="center" hgap="10" vgap="10">
+					<Button mnemonicParsing="false"
+						onAction="#handleCancelButtonClicked" text="%button-cancel-label" textAlignment="CENTER"
+						GridPane.columnIndex="0" GridPane.rowIndex="0" />
+					<Button mnemonicParsing="false" onAction="#handleApplyButtonClicked"
+						text="%button-apply-label" textAlignment="CENTER" GridPane.columnIndex="1"
+						GridPane.rowIndex="0" />
+				</GridPane>
+			</right>
+			<padding>
+				<Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+			</padding>
+			<BorderPane.margin>
+				<Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+			</BorderPane.margin>
+		</BorderPane>
+	</bottom>
+</BorderPane>
