+++ src/main/resources/fxml/CustomerEditView.fxml
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<BorderPane xmlns="http://javafx.com/javafx"
+	xmlns:fx="http://javafx.com/fxml"
+	fx:controller="de.htwsaar.prog3.carrental.controller.CustomerEditViewController">
+	<top>
+		<BorderPane prefHeight="25.0" prefWidth="1442.0" BorderPane.alignment="CENTER">
+			<left>
+				<GridPane alignment="center" hgap="10" vgap="10">
+					<Label text="%customer-edit-title-label" GridPane.columnIndex="0"
+						GridPane.rowIndex="0" />
+				</GridPane>
+			</left>
+			<padding>
+				<Insets left="27" right="5.0" top="12.0" />
+			</padding>
+			<BorderPane.margin>
+				<Insets left="27" right="5.0" top="12.0" />
+			</BorderPane.margin>
+		</BorderPane>
+	</top>
+	<center>
+		<GridPane alignment="center" hgap="10" vgap="10">
+			<Label text="%customer-first-name-label" GridPane.columnIndex="0" GridPane.rowIndex="0" />
+			<TextField fx:id="firstNameTextField" GridPane.columnIndex="0" GridPane.rowIndex="1" />
+			<Label text="%customer-last-name-label" GridPane.columnIndex="1" GridPane.rowIndex="0" />
+			<TextField fx:id="lastNameTextField" GridPane.columnIndex="1" GridPane.rowIndex="1" />
+			<Label text="%customer-email-label" GridPane.columnIndex="0" GridPane.rowIndex="2" />
+			<TextField fx:id="emailAddressTextField" GridPane.columnIndex="0" GridPane.rowIndex="3" />
+			<Label text="%customer-phone-number-label" GridPane.columnIndex="1" GridPane.rowIndex="2" />
+			<TextField fx:id="phoneNumberTextField" GridPane.columnIndex="1" GridPane.rowIndex="3" />
+			<Label text="%customer-date-of-birth-label" GridPane.columnIndex="2" GridPane.rowIndex="2" />
+			<TextField fx:id="dateOfBirthTextField" GridPane.columnIndex="2" GridPane.rowIndex="3" />
+			<Label text="%customer-street-label" GridPane.columnIndex="0" GridPane.rowIndex="4" />
+			<TextField fx:id="streetTextField" GridPane.columnIndex="0" GridPane.rowIndex="5" />
+			<Label text="%customer-house-number-label" GridPane.columnIndex="1" GridPane.rowIndex="4" />
+			<TextField fx:id="houseNumberTextField" GridPane.columnIndex="1" GridPane.rowIndex="5" />
+			<Label text="%customer-city-label" GridPane.columnIndex="2" GridPane.rowIndex="4" />
+			<TextField fx:id="cityTextField" GridPane.columnIndex="2" GridPane.rowIndex="5" />
+			<Label text="%customer-zip-code-label" GridPane.columnIndex="3" GridPane.rowIndex="4" />
+			<TextField fx:id="zipCodeTextField" GridPane.columnIndex="3" GridPane.rowIndex="5" />
+			<Label text="%customer-id-number-label" GridPane.columnIndex="0" GridPane.rowIndex="6" />
+			<TextField fx:id="idNumberTextField" GridPane.columnIndex="0" GridPane.rowIndex="7" />
+			<Label text="%customer-driver-license-id-label" GridPane.columnIndex="1"
+				   GridPane.rowIndex="6" />
+			<TextField fx:id="driverLicenseIdTextField" GridPane.columnIndex="1"
+				GridPane.rowIndex="7" />
+		</GridPane>
+	</center>
+	<bottom>
+		<BorderPane prefHeight="25.0" prefWidth="1442.0" BorderPane.alignment="CENTER">
+			<right>
+				<GridPane alignment="center" hgap="10" vgap="10">
+					<Button mnemonicParsing="false"
+						onAction="#handleCancelButtonClicked" text="%button-cancel-label" textAlignment="CENTER"
+						GridPane.columnIndex="0" GridPane.rowIndex="0" />
+					<Button mnemonicParsing="false" onAction="#handleApplyButtonClicked"
+						text="%button-apply-label" textAlignment="CENTER" GridPane.columnIndex="1"
+						GridPane.rowIndex="0" />
+				</GridPane>
+			</right>
+			<padding>
+				<Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+			</padding>
+			<BorderPane.margin>
+				<Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+			</BorderPane.margin>
+		</BorderPane>
+	</bottom>
+</BorderPane>
