--- src/main/resources/fxml/EditEmployeeView.fxml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<BorderPane xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
-	fx:controller="de.htwsaar.prog3.carrental.controller.EditEmployeeViewController">
-	<top>
-		<BorderPane prefHeight="25.0" prefWidth="1442.0" BorderPane.alignment="CENTER">
-			<left>
-				<GridPane alignment="center" hgap="10" vgap="10">
-					<Label text="%employee-edit-title-label" GridPane.columnIndex="0" GridPane.rowIndex="0" />
-				</GridPane>
-			</left>
-			<padding>
-				<Insets left="29" right="5.0" top="12.0" />
-			</padding>
-			<BorderPane.margin>
-				<Insets left="29" right="5.0" top="12.0" />
-			</BorderPane.margin>
-		</BorderPane>
-	</top>
-	<center>
-		<GridPane alignment="center" hgap="10" vgap="10">
-			<Label text="%employee-first-name-label" GridPane.columnIndex="0" GridPane.rowIndex="0" />
-			<TextField fx:id="firstNameTextField" GridPane.columnIndex="0" GridPane.rowIndex="1" />
-			<Label text="%employee-last-name-label" GridPane.columnIndex="1" GridPane.rowIndex="0" />
-			<TextField fx:id="lastNameTextField" GridPane.columnIndex="1" GridPane.rowIndex="1" />
-			<Label text="%employee-position-label" GridPane.columnIndex="2" GridPane.rowIndex="0" />
-			<TextField fx:id="positionTextField" GridPane.columnIndex="2" GridPane.rowIndex="1" />
-		</GridPane>
-	</center>
-	<bottom>
-		<BorderPane prefHeight="25.0" prefWidth="1442.0" BorderPane.alignment="CENTER">
-			<right>
-				<GridPane alignment="center" hgap="10" vgap="10">
-					<Button mnemonicParsing="false"
-						onAction="#handleCancelButtonClicked" text="%button-cancel-label" textAlignment="CENTER"
-						GridPane.columnIndex="0" GridPane.rowIndex="0" />
-					<Button mnemonicParsing="false" onAction="#handleApplyButtonClicked"
-						text="%button-apply-label" textAlignment="CENTER" GridPane.columnIndex="1"
-						GridPane.rowIndex="0" />
-				</GridPane>
-			</right>
-			<padding>
-				<Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-			</padding>
-			<BorderPane.margin>
-				<Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-			</BorderPane.margin>
-		</BorderPane>
-	</bottom>
-</BorderPane>
